embark.el: initial scuffed setup (no marginalia/consult/etc)
Trying out embark as a standalone package. The idea of
context-dependent keymaps is very promising. However the builtin
keymaps are a mishmash of random commands, so I tried to make them
smaller and more to the point (WIP).

TODO better hint (maybe hydra-style with commands grouped into columns
by category)